INFIVERSE-HR-PLATFORM/backend/tools/requirements.txt
# BHIV HR Platform - Tools Requirements
# Enhanced resume extraction and data processing tools

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0

# Resume processing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Natural Language Processing
spacy>=3.7.0
nltk>=3.8.0
textblob>=0.17.0

# Machine Learning (optional but recommended)
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# Database connectivity
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Text processing and validation
regex>=2023.0.0
phonenumbers>=8.13.0
email-validator>=2.1.0

# Logging and monitoring
colorlog>=6.7.0
tqdm>=4.66.0

# File handling
pathlib2>=2.3.0
chardet>=5.2.0

# Optional AI enhancements
# transformers>=4.35.0  # Uncomment for advanced NLP
# torch>=2.1.0          # Uncomment for deep learning models

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0